modified:   UTIL/chemmech/scripts/bldit_chemmech.csh         -allow script to workout without cmaq config file         -changed logic to delete build directory         -allow execution without argument; in this case, default compiler is intel  	modified:   UTIL/chemmech/scripts/run_chemmech.csh         -allow INPDIR and OUTDIR to be set by external environment variable         -allow script to work without cmaq config file         -changed logic to delete build directory         -allow execution without argument; in this case, default compiler is intel         -add CLEAR setting to delete existing output. current setting is TRUE .	modified:   UTIL/chemmech/src/Makefile         -set default compiler to intel  	modified:   UTIL/create_ebi/scripts/bldrun_create_ebi.csh         -allow INPDIR and OUTDIR to be set by external environment variable         -allow script to work without cmaq config file         -changed logic to delete build directory         -allow execution without argument; in this case, default compiler is intel         -add CLEAR setting to delete existing output. current setting is TRUE  	renamed:    UTIL/create_ebi/src/src_RXNSU -> UTIL/create_ebi/src  	renamed:    UTIL/create_ebi/src/degrade_codes_serial-RXNST -> UTIL/create_ebi/degrade_codes_serial-RXNST  	renamed:    UTIL/create_ebi/src/template_RXNSU_OPT -> UTIL/create_ebi/template_RXNSU_OPT         -moved directory's to the main developer's wishes  	modified:   UTIL/create_ebi/src/Makefile         -set default compiler to intel 	modified:   UTIL/inline_phot_preproc/scripts/bldrun.inline_phot_preproc.csh         -allow INPDIR and OUTDIR to be set by external environment variable         -allow script to work without cmaq config file         -changed logic to delete build directory         -allow execution without argument; in this case, default compiler is intel         -add CLEAR setting to delete existing output. current setting is TRUE         -sets cloud ice optics file by environment variables         -added check on log file for successful execution  	modified:   UTIL/inline_phot_preproc/src/optics_ice_cloud.f         -made opened file determined by environment variables  	modified:   UTIL/inline_phot_preproc/src/driver.f         -change codes for successful execution  	modified:   UTIL/inline_phot_preproc/src/inline_phot_preproc.makefile         -set default compiler to intel         -changed name of GC_INC  	renamed:    UTIL/inline_phot_preproc/src/flux_data -> UTIL/inline_phot_preproc/flux_data  	renamed:    UTIL/inline_phot_preproc/src/ice_clouds-> UTIL/inline_phot_preproc/ice_clouds  	renamed:    UTIL/inline_phot_preproc/src/photolysis_CSQY_data -> UTIL/inline_phot_preproc/photolysis_CSQY_data  	renamed:    UTIL/inline_phot_preproc/src/raw_alebo_data -> UTIL/inline_phot_preproc/raw_alebo_data  	renamed:    UTIL/inline_phot_preproc/src/water_clouds -> UTIL/inline_phot_preproc/refractive_indices/OPAC_water_clouds  	renamed:    UTIL/inline_phot_preproc/src/adient_aerosol_refrac_indx -> UTIL/inline_phot_preproc/refractive_indices/adient_aerosol_refrac_indx  	renamed:    UTIL/inline_phot_preproc/src/ice_clouds/ice_refractive_index.dat -> UTIL/inline_phot_preproc/refractive_indices/ice_refractive_index.dat  	renamed:    UTIL/inline_phot_preproc/src/water_clouds/water_refractive_index.dat -> UTIL/inline_phot_preproc/refractive_indices/water_refractive_index.dat         -moved directory's to the main developer's wishes. Also, ensure compatible to JPROC run-script.

+++ UTIL/inline_phot_preproc/scripts/bldrun.inline_phot_preproc.csh
@@ -28,17 +28,33 @@
    setenv compiler $1
    setenv compilerVrsn $2
  else
+   setenv compiler intel
+   setenv compilerVrsn Empty
+   echo "compiler and version not set"
    echo "usage: $0 <compiler>"
-   echo " where <compiler> is intel, pgi or gcc"
-   exit(2)
+   echo "setting compiler to intel"
  endif
 
+set echo 
 #> Source the config.cmaq file to set the build environment
- cd ../../..
- source ./config_cmaq.csh
+ if( -e ../../../config_cmaq.csh )then
+    cd ../../..
+    source ./config_cmaq.csh
+ else
+#work offline from CMAQ repository and build environment
+    setenv offline "Y"
+    setenv compilerString ${compiler}
+    setenv CMAQ_HOME $cwd/..
+ endif
 
 #> Source Code Repository
- setenv REPOROOT ${CMAQ_REPO}/UTIL/inline_phot_preproc  #> location of the source code for CSQY_TABLE_PROCESSOR
+ if( ! ( $?offline ) )then
+   setenv REPOROOT ${CMAQ_REPO}/UTIL/inline_phot_preproc  #> location of the source code for CHEMMECH
+ else
+   setenv REPOROOT ${CMAQ_HOME}
+ endif
+
+unset echo
 
 #===============================================================================
 #> Begin User Input Section 
@@ -48,49 +64,56 @@
  if ( ! $?MECH ) then
    set MECH =     'cb6r3_ae7_aq'
  endif
+ setenv CLEAR "TRUE" #> over-write existing output files
 
- set VRSN =      v532                       #> model version
+ if( ! ( $?offline ) )then
+   set WORKDIR = ${CMAQ_HOME}/UTIL/inline_phot_preproc
+ else
+   set WORKDIR = ${CMAQ_HOME}
+ endif
+ if ( ! $?INPDIR ) then
+    setenv INPDIR   ${WORKDIR}/input/${MECH}
+ endif
+ if ( ! $?OUTDIR ) then
+    setenv OUTDIR   ${WORKDIR}/output/${MECH}
+ endif
+
+ set VRSN =      v532                                 #> model version
  setenv EXEC     CSQY_TABLE_PROCESSOR_${VRSN}.exe     #> executable name for this application
- setenv WORKDIR  ${CMAQ_HOME}/UTIL/inline_phot_preproc
- setenv WORKREPO ${CMAQ_REPO}/UTIL/inline_phot_preproc
+ setenv WORKREPO ${REPOROOT}
  setenv BLDIR    ${WORKDIR}/scripts/BLD_CSQY_TABLE_${VRSN}_${compilerString}
- setenv INPDIR   ${WORKDIR}/input/${MECH}
- setenv OUTDIR   ${WORKDIR}/output/${MECH}
-
 #============================================================================================
 #> Set locations for source code and templates
 #============================================================================================
 
  set    SRCDIR = ${WORKREPO}/src
- setenv GC_INC ${CMAQ_REPO}/CCTM/src/MECHS/$MECH
 
 # Define environment variable for path to data module for photochemical mechanism
 # RXNS_DATA is the input directory containing the mechanism's data module
 # value will change based on user's goals. If the file is not found, this script
 # will check the output for CHEMMECH, and then check the CMAQ_REPO. If it is in 
 # neither of those places, the script aborts.
- if ( ! -e ${INPDIR} ) then
-    mkdir -p ${INPDIR}
- endif
 
 #> use RXNS_DATA_MODULE, comment out if CMAQ v5.02 and keep if CMAQ v5.1 or higher
  setenv USE_RXNS_MODULES T
  if( ${USE_RXNS_MODULES} == "T" )then
     setenv RXNS_DATA_SRC   ${INPDIR}/RXNS_DATA_MODULE.F90
-    if ( ! ( -e ${RXNS_DATA_SRC} ) )then
+    if ( ! ( -e ${RXNS_DATA_SRC} ) && ! ( $?offline ) )then
+       if ( ! -e ${INPDIR} ) then
+          mkdir -p ${INPDIR}
+       endif
        echo 'Cannot find RXNS_DATA_MODULE. Look in CHEMMECH Output...'
        if ( ! ( -e ${CMAQ_HOME}/UTIL/chemmech/output/${MECH}/RXNS_DATA_MODULE.F90 ) ) then
           echo 'RXNS_DATA_MODULE not in CHEMMECH Output. Look in CMAQ Repo...'
-          if ( ! -e ${CMAQ_REPO}/CCTM/src/MECHS/${MECH}/RXNS_DATA_MODULE.F90 ) then
-             echo 'RXNS_DATA_MODULE input file does not exist'
-             ls ${RXNS_DATA_SRC}
-             exit 1
-          else
-             cp ${CMAQ_REPO}/CCTM/src/MECHS/${MECH}/RXNS_DATA_MODULE.F90 $RXNS_DATA_SRC
-          endif
+          cp ${CMAQ_REPO}/CCTM/src/MECHS/${MECH}/RXNS_DATA_MODULE.F90 $RXNS_DATA_SRC
        else
           cp ${CMAQ_HOME}/UTIL/chemmech/output/${MECH}/RXNS_DATA_MODULE.F90 $RXNS_DATA_SRC
        endif
+    endif
+    if ( ! -e ${RXNS_DATA_SRC} ) then
+       echo 'Below RXNS_DATA_MODULE input file does not exist'
+       echo ${RXNS_DATA_SRC}
+       exit 1
     endif
  endif 
 
@@ -98,14 +121,11 @@
 #> Copy CSQY_TABLE_PROCESSOR Source Code into new build folder and compile
 #============================================================================================
  
- if ( ! -e "$BLDIR" ) then
-    mkdir -pv $BLDIR
- else
-    if ( ! -d "$BLDIR" ) then
-       echo "   *** target exists, but not a directory ***"
-       exit 1
-    endif
+ if ( -e "$BLDIR" ) then
+    echo "   *** build directory exist, deleting it***"
+    \rm -rf $BLDIR
  endif
+ mkdir -pv $BLDIR
  
 #> Whether to include spectral values of refractive indices for aerosol species [T|Y|F|N]
 #>  set F if CMAQ v5.02 and T if CMAQ v5.1 or higher
@@ -122,7 +142,7 @@
 setenv N_WAVEBANDS_OUT 7
 
 #>Compile the Executable
- cp -r ${SRCDIR}/* ${BLDIR}
+ cp -r ${SRCDIR}/* ${BLDIR}/.
  cd $BLDIR ; make clean; make -f inline_phot_preproc.makefile
  if( ! ( -e ${EXEC} ) )then
     echo "failed to compile ${BLDIR}/${EXEC}"
@@ -130,14 +150,26 @@ setenv N_WAVEBANDS_OUT 7
  endif
 
 #set up input data file directories
- set CSQY_DIR    = ${BLDIR}/photolysis_CSQY_data
- set REFRACT_DIR = ${BLDIR}/water_clouds
- set WVBIN_DIR   = ${BLDIR}/flux_data
+ set CSQY_DIR       = ${WORKREPO}/photolysis_CSQY_data
+ set REFRACT_DIR    = ${WORKREPO}/refractive_indices
+ set WVBIN_DIR      = ${WORKREPO}/flux_data
+ set ICE_CLOUDS_DIR = ${WORKREPO}/ice_clouds
+
+ set data_paths = ( ${CSQY_DIR} ${REFRACT_DIR} ${WVBIN_DIR} ${ICE_CLOUDS_DIR} )
+ foreach data_dir ( ${data_paths} )
+    if( ! ( -e ${data_dir} ) )cp -r $data_dir ${WORKDIR}/.
+ end
 
 # Define environment variables for inputs
- setenv GC_INC        ${INPDIR}
-#wavelength bin mapping data file
+#Wavelength bin mapping data file
  setenv WVBIN_FILE    $WVBIN_DIR/wavel-bins.dat
+
+#Files describing optical properties of ice cloud particles
+ setenv ICE_CLD_SSA ${ICE_CLOUDS_DIR}/fu96.ssa
+ setenv ICE_CLD_EXT ${ICE_CLOUDS_DIR}/fu96.ext
+ setenv ICE_CLD_ASY ${ICE_CLOUDS_DIR}/fu96.asy
+ setenv ICE_CLD_DEL ${ICE_CLOUDS_DIR}/fu96.del
+
 #Solar flux spectrum data file
  setenv FLUX_FILE     $WVBIN_DIR/solar-p05nm-UCI.dat
 #Raw cross-section and quantum yield data for photolysis rates
@@ -157,20 +189,39 @@ setenv N_WAVEBANDS_OUT 7
 #Set environment variables for the paths to each refractive index in
 #AE_REFRAC_LIST 
  setenv WATER     $REFRACT_DIR/water_refractive_index.dat
- setenv DUST      $REFRACT_DIR/inso00                    
- setenv SOLUTE    $REFRACT_DIR/waso00                    
- setenv SOOT      $REFRACT_DIR/soot00-two_way-Oct_21_2012
- setenv SEASALT   $REFRACT_DIR/ssam00                    
+ setenv DUST      $REFRACT_DIR/OPAC_water_clouds/inso00                    
+ setenv SOLUTE    $REFRACT_DIR/OPAC_water_clouds/waso00                    
+ setenv SOOT      $REFRACT_DIR/OPAC_water_clouds/soot00-two_way-Oct_21_2012
+ setenv SEASALT   $REFRACT_DIR/OPAC_water_clouds/ssam00                    
 
 #Define output directory variable and create
- if( ! ( -d $OUTDIR ) ) mkdir -p $OUTDIR
+ if( ( -d $OUTDIR ) )then
+    if( -e ${OUTDIR}/CSQY_DATA_${MECH} && -e ${OUTDIR}/PHOT_OPTICS.dat )then
+       if( $CLEAR == "FALSE")then
+         echo "Previous output exists; set CLEAR to TRUE to delete"
+         exit(1)
+       endif
+       \rm -rf $OUTDIR
+    endif
+ endif
+ mkdir -p $OUTDIR
  setenv OUT_DIR       $OUTDIR
-
+ 
 # Execute CSQY_TABLE_PROCESSOR
- $BLDIR/$EXEC >&! bldrun.log
+#cd ${WORKDIR}
+ $BLDIR/$EXEC >&! ${WORKDIR}/scripts/bldrun.log
+ set signal = ` tail -1 ${WORKDIR}/scripts/bldrun.log `
+echo " "
+ echo ${signal}
+ set output = $signal # `grep "NORMAL_STOP" bldrun.log `
+ echo $output
+echo " "
 
- if ( $? != 0 ) then
+#cd ${WORKDIR}/scripts
+
+ if ( $? != 0  || $output != "NORMAL_STOP" ) then
     echo "INLINE_PHOT_PREPROC ($BLDIR/$EXEC) failed for some reason. Halt Build Process!"
+    echo "Try checking end of ${WORKDIR}/scripts/bldrun.log"
     exit 1
  endif
 
@@ -178,7 +229,12 @@ setenv N_WAVEBANDS_OUT 7
 
 echo " "
 echo " "
-echo "Check directory ${OUTDIR} for CSQY_DATA_${MECH} and PHOT_OPTICS.dat files"
+if( ( -e ${OUTDIR}/CSQY_DATA_${MECH} ) && -e ${OUTDIR}/PHOT_OPTICS.dat )then
+  echo "CSQY_DATA_${MECH} and PHOT_OPTICS.dat files created. Check each in ${OUTDIR}"
+else
+  echo "CSQY_DATA_${MECH} and PHOT_OPTICS.dat files not created."
+  echo "Check end of ${WORKDIR}/scripts/bldrun.log"
+endif
 echo " "
 echo " "
